AI_CRYPTO swing : exécution réelle des trades via Alpaca paper (désactivée par défaut)
Répond à la demande explicite d'essayer des trades pour de vrai (paper,
compte Alpaca partagé déjà utilisé par WATCHLIST/AI_BLUECHIP/AI_PENNY).

- alpaca_data.py::submit_crypto_market_order -- copie de submit_market_
  order avec les 2 différences obligatoires côté crypto : qty
  fractionnaire (float, pas int) et TimeInForce.GTC (DAY est rejeté par
  l'API Alpaca pour le marché crypto 24/7). _to_alpaca_crypto_symbol
  convertit 'BTC-USD' (format interne) -> 'BTC/USD' (format Alpaca).
  Vérifié en direct : crypto_status du compte paper = ACTIVE, BTC/USD et
  ETH/USD tradables et fractionable.

- tasks.py::_execute_alpaca_paper_trades_for_crypto_swing +
  crypto_swing_paper_trade_task -- utilise scan_crypto_swing (pas scan_
  crypto_drip, jamais câblé). Désactivé par défaut (AI_CRYPTO_SWING_
  ENABLED='false', contrairement à AI_CRYPTO_ENABLED dont le défaut
  'true' n'a jamais eu de conséquence faute d'être appelé) -- active
  explicitement de vrais ordres papier dès activation. Position sizing
  volontairement prudent (5% par défaut, stratégie non prouvée). Aucun
  PaperTrade créé si l'ordre Alpaca échoue réellement. Circuit breaker
  dédié (broker_path='ALPACA_SWING', ajouté à reset_daily_equity_
  breaker). Planifiée toutes les 4h (marché crypto 24/7, décisions sur
  bougies journalières -- pas besoin de vérifier plus souvent).

- BUG TROUVÉ EN CONSTRUISANT LES TESTS : PaperTrade.quantity était un
  IntegerField -- tronquait silencieusement tout qty<1 à 0. À 50 000$ le
  BTC, une position normale (5% d'un compte 100k$) donne 0.1 BTC, pas 1 :
  aurait annulé silencieusement CHAQUE position crypto correctement
  dimensionnée. Touchait aussi le chemin SIM existant (_execute_paper_
  trades_for_crypto_sandbox, qty = int(...)), jamais surfacé faute d'être
  appelé. Migration 0037 : IntegerField -> FloatField (sans perte pour
  les sandboxes actions existantes).

- Tentative d'amélioration ML supplémentaire (HistGradientBoosting au
  lieu de RandomForest, testée en direct sur BTC/ETH 14j/5%) : walk-
  forward F1 quasi identique (0.290 vs 0.301) -- pas assez d'amélioration
  pour justifier le changement, RandomForest conservé.

Tests : 8 (tests_alpaca_crypto.py, symbole + GTC/qty fractionnaire) + 12
(tests_crypto_swing_alpaca_execution.py, désactivé par défaut, pas de
compte, circuit breaker, entrée/sortie réelles, ordre échoué = pas de
PaperTrade fantôme, doublon évité, watchlist DB prioritaire). 175 tests
portfolio+analysis, tous verts.

+++ portfolio_backend/portfolio/models.py
@@ -107,7 +107,15 @@ class PaperTrade(models.Model):
 	ticker = models.CharField(max_length=10)
 	sandbox = models.CharField(max_length=20, choices=SANDBOX_CHOICES, default='WATCHLIST', db_index=True)
 	entry_price = models.DecimalField(max_digits=10, decimal_places=2)
-	quantity = models.IntegerField()
+	# IntegerField -> FloatField le 2026-08-18 : le chemin crypto (SIM
+	# existant ET le nouveau chemin Alpaca swing) a besoin de quantités
+	# fractionnaires (ex. 0.1 BTC) -- un IntegerField tronque silencieusement
+	# tout qty<1 à 0, ce qui aurait annulé silencieusement toute position
+	# crypto dimensionnée normalement (5% d'un compte 100k$ / 50000$ le BTC
+	# = 0.1 BTC, pas 0). Compatible sans perte avec les sandboxes actions
+	# existantes (quantités entières déjà stockées restent des floats
+	# équivalents, ex. 100 -> 100.0).
+	quantity = models.FloatField()
 	entry_date = models.DateTimeField(auto_now_add=True)
 	entry_signal = models.FloatField(null=True, blank=True)
 	entry_features = models.JSONField(null=True, blank=True)
